QX-3771 - Merge v3.1.81 into mc master simple (#31)
We found that the issues with page orientation during printing (QX-3771) have been resolved in the latest releases of Mozilla's PDFjs, which we've been meaning to pull in anyway.

This PR is to merge Mozilla's v3.1.81 into the mc-master branch. In doing so we had to resolve a lot of conflicts and changes primarily around styling. Hopefully we've done so in a way to make future merges from Mozilla's repo a bit easier... but know there will still be conflicts. Perhaps in the future we can solve that a different way.